Separate execution tests into `EndToEndTests` target.
