+++ Sources/ValCommand/String+Log.swift
@@ -1,5 +1,3 @@
-import ValCommand
-
 extension String: Log {
 
   public var hasANSIColorSupport: Bool { false }
