+++ Tests/TestSupport/FileManager+Extension.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+extension FileManager {
+
+  /// Returns the URL of a temporary file.
+  public func temporaryFile() -> URL {
+    temporaryDirectory.appendingPathComponent("\(UUID())")
+  }
+
+}
